# Use inline storage for the common case of 1 STREAM frame per packet