--- a/quinn-proto/Cargo.toml
+++ b/quinn-proto/Cargo.toml
@@ -34,6 +34,7 @@ rustls = { version = "0.19", features = ["quic"], optional = true }
 rustls-native-certs = { version = "0.5", optional = true }
 slab = "0.4"
 thiserror = "1.0.21"
+tinyvec = { version = "1.1", features = ["alloc"] }
 tracing = "0.1.10"
 webpki = { version = "0.21", optional = true }