When packets are transmitted, the most common case them is to contain at
most a single STREAM frame. For this common case the current system does
still 2 allocations: One for the `SentPacket` structure which tracks the whole
packet, and one for the stream metadata. This double allocation places
some additional pressure on the allocator. This change changes
`SentPacket::stream_frames`  to make use of vector with small element
optimization, in order to be able to handle the common case of sending
1 stream frame without any further allocation.

Benchmark results
===

Overall the differences don't seem huge. However they become bigger when other
optimizations (like bigger MTUs and size reduction for `SentPacket` are also
enabled).

Without GSO
--

**Before:**

> Sent 4294967296 bytes on 1 streams in 15.37s (266.49 MiB/s)

**After:**

> Sent 4294967296 bytes on 1 streams in 15.33s (267.17 MiB/s)

With GSO
--

**Before:**

> Sent 4294967296 bytes on 1 streams in 11.79s (347.48 MiB/s)

**After:**

> Sent 4294967296 bytes on 1 streams in 11.70s (350.11 MiB/s)