@Ralith 2021-02-08
I'm on board with this change, and I suspect we can make a lot of others like it to start adding up to a larger impact. Tinyvec definitely could be used here and this, if any, is an application where taking care with `unsafe` is justified, though, so we should give that a try.