@Matthias247 2021-02-08
First time I heard about the `SmallVec` issues, despite having it used in some other projects. Interesting that those are indeed legitimate bugs, instead of more debatable UB discussions where things work fine in practice.

I'm not terribly concerned by the past issues, since past issues always mean people took a closer look and fixed them. The crate is fairly widely used, and so should have a large level of maturity. Besides that, the set of APIs that are used here are also fairly minimal and unlikely to break.

I looked a bit at the difference with `TinyVec`, and it seems like the main drawbacks of those is that it default initializes all elements, and only supports `Default` types. I think for this particular use-case it won't matter a lot. For other potential use-cases (e.g. optimizing `RangeSet`) the `SmallVec` benefits could however be more important, and sticking with a single dependency might be preferable.

Overall I'm fine with picking any of those for the moment.