@djc 2021-02-08
> I'm not terribly concerned by the past issues, since past issues always mean people took a closer look and fixed them. The crate is fairly widely used, and so should have a large level of maturity.

Usually I also think this way, but given this track record:

- Jan 8, 2021
- Jul 19, 2019
- Jun 6, 2019
- Sep 25, 2018
- Jul 19, 2018

I see little proof that people started to look closer, nor that wide usage or a high maturity level have helped here.