@djc 2021-02-08
(If we truly need a minimal subset of API, maybe we should NIH it with a little enum?)