@Ralith 2021-02-08
> For other potential use-cases (e.g. optimizing RangeSet) the SmallVec benefits could however be more important

Can you elaborate here? `Range<u64>` is easy enough to initialize, after all.