@Matthias247 2021-02-08
This was more about a longer Inline Array length potentially making sense there