@Ralith 2021-02-08
I'm confused, then. Tinyvec should still provide that benefit, no? Is there any case we have that smallvec should be expected to perform better on?