@Matthias247 2021-02-08
My understanding is that TinyVec initialed all elements whereas SmallVec leaves them Uninit until use. So the creation costs are a tiny bit higher