@Matthias247 2021-02-09
Updated for TinyVec. We can revisit once the other optimizations have landed.