@Matthias247 2021-02-09
Seems like `Default` isn't implemented for `Range` on 1.45.

Not sure when that was added.
Options:
1. Increase minimum supported version
2. Back to SmallVec
3. Add a wrapper around Range

I'm not really a fan of 3. Happy with any of the other versions.

Regarding "NIH": I prefer to avoid it, since we then just have one more place where we discuss about unsafe code. I rather would spend the same time reviewing the utilized code path in SmallVec.