@djc 2021-02-09
Other option: implement custom `Default` for `StreamMeta`? Seems like the most attractive option to me.