@Matthias247 2021-02-10
> Some minor nits left. Did you observe a perf difference between smallvec and tinyvec?

I didn't really benchmark. I thought it would first by worthwhile to improve a few other code parts, and then do some comparisons which simply swap those libraries and check the overall result.